requirements.txt
tqdm
hydra-core
einops
pytorch_lightning
matplotlib
joblib
mido
pydub
git+https://github.com/KinWaiCheuk/AudioLoader.git
torch
torchvision
torchaudio
scikit-learn
mir_eval